refactor: Rename package to among_them
