+++ pyproject.toml
@@ -1,10 +1,13 @@
 [tool.poetry]
-name = "llm_postor"
+name = "among_them"
 version = "0.1.0"
 description = "This project tests and compares the persuasive abilities of large language models (LLMs) in a game-like environment simulating 'Among Us', where AI agents powered by LLMs interact, make decisions, and attempt to persuade each other."
 authors = ["Marcin Korcz", "Andrii Chmutov", "Vasyl Korzavatykh", "Mateusz Idziejczak", "Mateusz Stawicki"]
 readme = "README.md"
 package-mode = true
+packages = [
+    { include = "among_them", from = "src" }
+]
 
 [tool.poetry.dependencies]
 python = "^3.11"
