+++ scripts/tournament.py
@@ -7,7 +7,7 @@ def main():
     # (this assumes the script is located in the scripts/ folder)
     project_root = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
     os.chdir(project_root)
-    tournament_path = os.path.join(project_root, "src", "llm_postor", "tournament.py")
+    tournament_path = os.path.join(project_root, "src", "among_them", "tournament.py")
 
     if not os.path.exists(tournament_path):
         raise FileNotFoundError(f"File does not exist: {tournament_path}")
