+++ src/among_them/config.py
@@ -14,8 +14,8 @@ if not OPENROUTER_API_KEY:
         "in your environment or in a .env file in the project root."
     )
 
-# if src/llm_postor/game/dummy.py does not exist, create it.
+# if src/among_them/game/dummy.py does not exist, create it.
 # This is for abusing streamlit refresh when game_state.json changes
-if not os.path.exists("src/llm_postor/game/dummy.py"):
-    with open("src/llm_postor/game/dummy.py", "w") as f:
+if not os.path.exists("src/among_them/game/dummy.py"):
+    with open("src/among_them/game/dummy.py", "w") as f:
         f.write("timestamp = '2024-11-15 00:20:17.946790'")
