+++ src/among_them/game/agents/adventure_agent.py
@@ -4,15 +4,15 @@ from langchain.schema import HumanMessage, SystemMessage
 from langchain_openai import ChatOpenAI
 from pydantic import Field
 
-from llm_postor.config import OPENROUTER_API_KEY
-from llm_postor.game.agents.base_agent import Agent
-from llm_postor.game.llm_prompts import (
+from among_them.config import OPENROUTER_API_KEY
+from among_them.game.agents.base_agent import Agent
+from among_them.game.llm_prompts import (
     ADVENTURE_ACTION_SYSTEM_PROMPT,
     ADVENTURE_ACTION_USER_PROMPT,
     ADVENTURE_PLAN_SYSTEM_PROMPT,
     ADVENTURE_PLAN_USER_PROMPT,
 )
-from llm_postor.game.utils import check_action_valid
+from among_them.game.utils import check_action_valid
 
 
 class AdventureAgent(Agent):
