+++ src/among_them/game/agents/discussion_agent.py
@@ -4,8 +4,8 @@ from langchain.schema import HumanMessage, SystemMessage
 from langchain_openai import ChatOpenAI
 from pydantic import Field
 
-from llm_postor.config import OPENROUTER_API_KEY
-from llm_postor.game.llm_prompts import (
+from among_them.config import OPENROUTER_API_KEY
+from among_them.game.llm_prompts import (
     DISCUSSION_RESPONSE_SYSTEM_PROMPT,
     DISCUSSION_RESPONSE_USER_PROMPT,
     DISCUSSION_SYSTEM_PROMPT,
