+++ src/among_them/game/chat_analyzer.py
@@ -5,8 +5,8 @@ from langchain.schema import HumanMessage, SystemMessage
 from langchain_openai import ChatOpenAI
 from pydantic import BaseModel
 
-from llm_postor.game.llm_prompts import ANNOTATION_SYSTEM_PROMPT
-from llm_postor.game.players.base_player import Player
+from among_them.game.llm_prompts import ANNOTATION_SYSTEM_PROMPT
+from among_them.game.players.base_player import Player
 
 
 class ChatAnalyzer(BaseModel):
