+++ src/among_them/game/models/history.py
@@ -3,9 +3,9 @@ from typing import List
 
 from pydantic import BaseModel, Field
 
-from llm_postor.game.models.engine import GameLocation, GamePhase
-from llm_postor.game.models.tasks import Task
-from llm_postor.game.models.usage_metadata import UsageMetadata
+from among_them.game.models.engine import GameLocation, GamePhase
+from among_them.game.models.tasks import Task
+from among_them.game.models.usage_metadata import UsageMetadata
 
 
 class PlayerState(str, Enum):
