+++ src/among_them/game/players/fake_ai.py
@@ -1,7 +1,7 @@
 import random
 from typing import List
 
-from llm_postor.game.players.base_player import Player
+from among_them.game.players.base_player import Player
 
 
 class FakeAIPlayer(Player):
